\name{genotypes}
\alias{genotypes}
\title{Genotype data from Aleppo Pines}
\description{
  Multiple sampled alleles from Aleppo Pine (Pinus halepensis) in the Eastern Mediterranean. Adapted from Gershberg et al. 2016.
}
\usage{
genotypes
}
\source{
Gershberg, A., Ne'eman, G., Ben-Shlomo, R. (2016). Genetic structure of a
naturally regenerating post-fire seedling population: Pinus halepensis as
a case study. Frontiers in Plant Science 7: 549.
}

\keyword{file}
