\name{moth}
\alias{moth}
\title{Temporal allele frequency shifts}
\description{
  Change in allele frequency over eight generations in the scarlet tiger moth (Callimorpha
dominula).
}
\usage{
moth
}
\source{
Fisher, R. A. and E. B. Ford (1947). The Spread of a Gene in Natural Conditions
in a Colony of the Moth Panaxia dominula. Heredity 1: 143-174.
}

\keyword{file}
