\name{snp}
\alias{snp}
\title{SNP information for 25 loci}
\description{
  Single Nucleotide Polymorphism data for 25 loci from the 1,000 Genomes Project (http://www.internationalgenome.org). Includes information on allele frequency, homozygosity, heterozygosity, chromosome location, and functional location of SNP.
}
\usage{
snp
}
\source{
Auton, A., Abecasis, G., Altshuler, D. et al. (2015) "A global reference for human genetic variation" <doi:10.1038/nature15393>
}

\keyword{file}
