\name{thirteen}
\alias{thirteen}
\title{Genotypes across 13 CODIS loci}
\description{
Genotypes of 1,036 individuals across 13 core CODIS (Combined DNA Index System) loci, collected in the USA.
}
\usage{
thirteen
}
\source{
Hill, C.R., Duewer, D.L., Kline, M.C., Coble, M.D., Butler, J.M. (2013) "U.S. population data for 29 autosomal STR loci" <doi:10.1016/j.fsigen.2012.12.004>
}

\keyword{file}
