% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnoRan.R
\name{AnnoSin}
\alias{AnnoSin}
\title{AnnoSin: Single-Position Annotation Based on Genomic Annotations}
\usage{
AnnoSin(data = data, annotation = annotation)
}
\arguments{
\item{data}{A data frame containing the genomic positions to be annotated. This should include columns for 
chromosome (`CHR`) and position (`POSITION`).}

\item{annotation}{A data frame containing gene annotations. The annotation data frame must include columns 
for `chromosome_name`, `start_position`, `end_position`, and `external_gene_name`.}
}
\value{
A data frame that includes the original positions along with an additional column, `external_gene_name`, 
containing the gene name corresponding to each position.
}
\description{
This function annotates individual genomic positions within a dataset by matching them with corresponding gene 
annotations. It takes a data frame with genomic positions and annotates them using an external annotation file 
or data frame.
}
\details{
The function uses the `POSITION` and `CHR` columns from the input `data` to find the corresponding 
gene name in the `annotation` data frame. For each position, it checks whether the position lies within the 
start and end coordinates of a gene in the specified chromosome. If a match is found, the corresponding gene 
name is assigned; otherwise, `NA` is returned.
}
\note{
The function expects the `annotation` data frame to contain specific columns: `chromosome_name`, 
`start_position`, `end_position`, and `external_gene_name`.
}
\examples{
# Example usage:
positions_data <- data.frame(CHR = c("1", "2"), POSITION = c(1500, 6000))
annotation_df <- data.frame(chromosome_name = c("1", "2"), start_position = c(1000, 5000), 
                            end_position = c(2000, 6500), external_gene_name = c("GeneA", "GeneB"))
annotated_data <- AnnoSin(data = positions_data, annotation = annotation_df)

}
