\name{fly}
\alias{fly}
\title{Drosphila melanogaster bw75 data}
\description{
  A matrix of observed allele counts from Buri (1956). Each row contains observations starting at generation 1 in the first row, across 107 replicates of the number of bw75 allele.
}

\usage{
fly
}

\source{
Buri, P. (1956). "Gene frequency in small populations of mutant Drosophila" <doi:10.1111/j.1558-5646.1956.tb02864.x>
}

\keyword{file}
