\name{genotypes}
\alias{genotypes}
\title{Genotype data from Aleppo Pines}
\description{
  Multiple sampled alleles from Aleppo Pine (Pinus halepensis) in the Eastern Mediterranean. Adapted from Gershberg et al. 2016.
}
\usage{
genotypes
}
\source{
Gershberg, A., Ne'eman, G., Ben-Shlomo, R. (2016) "Genetic structure of a naturally regenerating post-fire seedling population: Pinus halepensis as a case study" <doi:10.3389/fpls.2016.00549>
}

\keyword{file}
