\name{moth}
\alias{moth}
\title{Temporal allele frequency shifts}
\description{
  Change in allele frequency over eight generations in the scarlet tiger moth (Callimorpha
dominula).
}
\usage{
moth
}
\source{
Fisher, R.A. and E.B. Ford (1947) "The Spread of a Gene in Natural Conditions in a Colony of the Moth Panaxia dominula" <doi:10.1038/hdy.1947.11>
}

\keyword{file}
