% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plQC.R
\name{plQC}
\alias{plQC}
\title{Perform Quality Control and PCA Analysis using PLINK}
\usage{
plQC(file = file, maf = 0.05, gen = 0.1, mind = 0.1, hwe = 1e-06, chrset = 22)
}
\arguments{
\item{file}{A character string specifying the path to the input file (without extension). The function accepts both 
`.ped` and `.bed` file formats.}

\item{maf}{A numeric value representing the minor allele frequency threshold for filtering. Default is 0.05.}

\item{gen}{A numeric value representing the genotype missingness threshold. Default is 0.1.}

\item{mind}{A numeric value representing the individual missingness threshold. Default is 0.1.}

\item{hwe}{A numeric value representing the p-value threshold for Hardy-Weinberg equilibrium test. Default is 1e-6.}

\item{chrset}{An integer representing the chromosome set to be analyzed. Default is 22.}
}
\value{
A PCA plot showing the first two principal components (PC1 and PC2) of the filtered genotype data.
}
\description{
This function performs quality control (QC) on genetic data using PLINK and calculates principal components 
for the provided genotype data. It generates a PCA plot based on the resulting principal components.
}
\details{
The function uses the PLINK tool to perform QC filtering on the input genetic data and calculates 
         the principal components. The PCA plot visualizes the distribution of the samples based on the first 
         two principal components, highlighting potential clusters or outliers in the data.
}
\note{
Ensure that the PLINK executable is accessible at the specified path in the `extdata/plink/` directory.
      The input files must be in the correct format, and the function assumes the presence of `ggplot2` for plotting.
}
\examples{
# Example usage:
plQC(file = "path/to/data", maf = 0.05, gen = 0.1, mind = 0.1, hwe = 1e-6, chrset = 22)

}
