% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essential-pro.R
\name{read.vcf}
\alias{read.vcf}
\title{read.vcf: Read and Clean VCF Files}
\usage{
read.vcf(file, special.char = "##", ...)
}
\arguments{
\item{file}{A character string representing the path to the VCF file to be read.}

\item{special.char}{A character string that indicates the type of comment lines to be removed from the VCF file. The default is `"##"` which is typical for VCF files.}

\item{...}{Additional arguments to be passed to the `read.table` function for reading the data.}
}
\value{
A data frame containing the cleaned VCF data, with appropriate column names for the genomic data.
}
\description{
This function reads a Variant Call Format (VCF) file, removes specific comment lines, and returns the data as a data frame.
}
\details{
The function reads all lines from the specified VCF file, removes lines starting with the specified comment character (defaulting to `##`), and replaces the header line containing `#CHROM` with `CHROM`. The cleaned lines are then converted into a data frame using `read.table`.
}
\note{
The input file must be a valid VCF file. The function assumes that the format of the file is correct and consistent with VCF standards.
}
\examples{
# Example usage:
vcf_data <- read.vcf("path/to/sample.vcf")

# Output: A data frame representing the cleaned VCF data

}
