\name{snp}
\alias{snp}
\title{SNP information for 25 loci}
\description{
  Single Nucleotide Polymorphism data for 25 loci. Includes information on allele frequency, homozygosity, heterozygosity, chromosome location, and functional location of SNP.
}
\usage{
snp
}
\source{
1,000 Genomes Project (http://www.internationalgenome.org)
}

\keyword{file}
