\name{thirteen}
\alias{thirteen}
\title{Genotypes across 13 CODIS loci}
\description{
Genotypes of 1,036 individuals across 13 core CODIS (Combined DNA Index System) loci, collected in the USA.
}
\usage{
thirteen
}
\source{
Hill, C.R., Duewer, D.L., Kline, M.C., Coble, M.D., and Butler, J.M. (2013) U.S. population data for 29 autosomal STR loci. Forensic Sci. Int. Genet. 7: e82-e83.
Accessed from: https://strbase.nist.gov/fbicore.htm
}

\keyword{file}
