\name{whale}
\alias{whale}
\title{Genotypes of 246 South Pacific Blue Whales}
\description{
Data frame of genotypes collected from 264 individual blue whales (Balaenoptera musculus) across seven distinct loci. Adapted from Attard et al. 2012.
}
\usage{
whale
}
\source{
Attard, C. R., Beheregaray, L. B., Jenner, K. C. S., Gill, P. C., Jenner, M. N., Morrice, M. G., Robertson, K.M. & Moller, L. M. (2012) 
"Hybridization of Southern Hemisphere blue whale subspecies and a sympatric area off Antarctica: impacts of whaling or climate change?"<doi:10.1111/mec.12025>
}

\keyword{file}
